requirements.txt
tradingview-ta
pandas
yfinance
requests
python-dotenv
pytz